file structure refactor for profile page
